[SOL-32] Replace entity rotations from normal to radians

+++ Assets/SolidSpace/Scripts/Entities/Components/Scripts/RotationComponent.cs
@@ -5,6 +5,6 @@ namespace SolidSpace.Entities.Components
 {
     public struct RotationComponent : IComponentData
     {
-        public half value;
+        public float value;
     }
 }
